refacter: change agencia to be smaller, add address class connected to agencia
