+++ src/main/java/com/ada/santander/coders/locadora/entity/Agencia.java
@@ -1,6 +1,5 @@
 package com.ada.santander.coders.locadora.entity;
 
-
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -15,18 +14,10 @@ public class Agencia {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    @Column(name = "CNPJ")
-    private String cnpj;
-
-    @Column(name = "RAZAO_SOCIAL")
-    private String razaoSocial;
-
-    @Column(name = "NOME_FANTASIA")
-    private String nomeFantasia;
-
     @Column(name = "TAMANHO_MAXIMO_DA_FROTA")
-    private Integer tamanhoMaximoDaFrota;
+    private int tamanhoMaximoDaFrota;
 
-    @Column(name = "ENDERECO")
-    private Integer endereco;
+    @OneToOne
+    @JoinColumn(name = "ENDERECO_ID", referencedColumnName = "id")
+    private Endereco endereco;
 }
