--- src/main/java/com/ada/santander/coders/locadora/model/Endereco.java
@@ -1,10 +1,0 @@
-package com.ada.santander.coders.locadora.model;
-
-public class Endereco {
-    private int cep;
-    private String logradouro;
-    private String bairro;
-    private String uf;
-    private String estado;
-    private String regiao;
-}
