+++ src/main/java/com/ada/santander/coders/locadora/repository/AgenciaRepository.java
@@ -5,9 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface AgenciaRepository extends JpaRepository<Agencia,Long> {
     Page<Agencia> findAll(Pageable pageable);
-    Optional<Agencia> findAgenciaByRazaoSocial(String razaoSocial);
 }
